refactor: extract common scraping and browser utilities
Create flows/common.py with shared upload_html_to_datalake task and move browser helpers (deny_consent_banner, scroll_to_top) to ingestion/browser.py.
Remove duplicate code from netto and rewe scrapers, update string formatting to use f-strings, and clean up unused imports.
