+++ backend/app/flows/scrape_netto_offers.py
@@ -1,4 +1,3 @@
-from datetime import date
 from itertools import batched
 from pathlib import Path
 from typing import Literal
@@ -6,7 +5,7 @@ from typing import Literal
 from prefect import flow, get_run_logger, task
 
 from core.config import get_prefect_settings
-from core.supabase import get_supabase
+from flows.common import upload_html_to_datalake
 from ingestion.browser import seleniumbase_browser_context
 from ingestion.offers.netto_scrapper import (
     StoreConfig,
@@ -23,15 +22,6 @@ settings = get_prefect_settings()
 STORE_CONFIGS_PATH = Path(__file__).parent / "../ingestion/offers/netto_store_configs.json"
 STORE_CONFIGS = get_store_configs(STORE_CONFIGS_PATH)
 
-DATALAKE_PATH_TEMPLATE = (
-    "retailer={retailer}/"
-    "store_external_id={store_external_id}/"
-    "year={year}/"
-    "month={month}/"
-    "day={day}/"
-    "{page_name}.html"
-)
-
 
 @task(retries=2, retry_delay_seconds=30)
 @forward_logs
@@ -42,35 +32,6 @@ def extract_html_with_seleniumbase(conf: StoreConfig) -> dict[Literal["main", "a
     with seleniumbase_browser_context() as ctx:
         page = ctx.new_page()
         return scrape_store(conf=conf, page=page)
-
-
-@task
-@forward_logs
-def upload_html_to_datalake(conf: StoreConfig, html_content: str, page_name: str) -> str:
-    logger = get_run_logger()
-    supabase = get_supabase()
-
-    today = date.today()
-    path = DATALAKE_PATH_TEMPLATE.format(
-        retailer=conf.retailer,
-        store_external_id=conf.external_id,
-        year=today.year,
-        month=today.strftime("%m"),
-        day=today.strftime("%d"),
-        page_name=page_name,
-    )
-
-    logger.info(f"Uploading page {page_name} to Supabase path: {path}")
-
-    supabase.storage.from_("raw_offers").upload(
-        path,
-        html_content.encode("utf-8"),
-        file_options={
-            "content-type": "text/html",
-            "upsert": "true",
-        },
-    )
-    return path
 
 
 @flow(name="scrape-netto-daily-offers")
