+++ backend/app/ingestion/offers/rewe_scrapper.py
@@ -1,6 +1,5 @@
 import functools
 import json
-import time
 from dataclasses import dataclass
 from datetime import date
 from pathlib import Path
@@ -10,7 +9,7 @@ from bs4 import BeautifulSoup
 from loguru import logger
 from playwright.sync_api import Page
 
-from ingestion.browser import browser_context
+from ingestion.browser import browser_context, deny_consent_banner, scroll_to_top
 
 
 @dataclass(frozen=True, slots=True)
@@ -39,7 +38,6 @@ STORE_CONFIGS: list[StoreConfig] = [
 
 DENY_BUTTON_CSS_LOCATOR = 'button[data-testid="uc-deny-all-button"]'
 DATE = date.today()
-ISODATE = DATE.isoformat()
 HTML_FILE_TEMPLATE = (
     "{root}/"
     "retailer={retailer}/"
@@ -73,20 +71,14 @@ def main(
         articles_file = file_path_fn(name="articles")
 
         if main_file.exists() and articles_file.exists():
-            logger.info("{} and {} already exist. Skipping ...", main_file, articles_file)
+            logger.info(f"{main_file} and {articles_file} already exist. Skipping ...")
             continue
         with browser_context(headless=headless) as ctx:
             page = ctx.new_page()
             html_content = scrape_store(url=conf.url, page=page)
 
-        save_scraped_html(
-            html=html_content["main"],
-            file_path=file_path_fn(name="main"),
-        )
-        save_scraped_html(
-            html=html_content["articles"],
-            file_path=file_path_fn(name="articles"),
-        )
+        save_scraped_html(html=html_content["main"], file_path=file_path_fn(name="main"))
+        save_scraped_html(html=html_content["articles"], file_path=file_path_fn(name="articles"))
 
 
 def get_file_path_with_retailer_store_date_partitions(
@@ -99,8 +91,7 @@ def get_file_path_with_retailer_store_date_partitions(
         external_id=conf.external_id,
         year=DATE.year,
         month=DATE.strftime("%m"),
-        day=DATE.strftime(r"%d"),
-
+        day=DATE.strftime("%d"),
     )
     return Path(path_str)
 
@@ -110,18 +101,18 @@ def scrape_store(
     page: Page,
 ) -> dict[Literal["main", "articles"], str]:
     page.goto(url)
-    deny_usercentrics_banner(page)
+    deny_consent_banner(page, DENY_BUTTON_CSS_LOCATOR)
 
     page.wait_for_timeout(1000)
 
     page.wait_for_selector("footer").scroll_into_view_if_needed()
-    scroll_to_top(page)  # prime all content before scrapping
+    scroll_to_top(page)
 
     main_locator = page.get_by_role("main").first
     main_soup = BeautifulSoup(main_locator.inner_html(), "html.parser")
     articles = main_soup.find_all("article")
 
-    logger.info("Found {} items", len(articles))
+    logger.info(f"Found {len(articles)} items")
     return {
         "main": main_soup.prettify(),
         "articles": "".join([a.prettify() for a in articles]),
@@ -129,55 +120,12 @@ def scrape_store(
 
 
 def save_scraped_html(html: str, file_path: Path) -> None:
-    logger.info("Saving to {}", file_path)
+    logger.info(f"Saving to {file_path}")
     if not (parent_dir := file_path.parent).exists():
-        logger.info("Creating folder {}", parent_dir)
+        logger.info(f"Creating folder {parent_dir}")
         parent_dir.mkdir(parents=True, exist_ok=True)
     with file_path.open("wt") as f:
         f.write(html)
-
-
-def deny_usercentrics_banner(page) -> None:
-    try:
-        # 1. Locate the button directly. Playwright pierces the Shadow DOM automatically.
-        deny_button = page.locator(DENY_BUTTON_CSS_LOCATOR)
-
-        # 2. Wait up to 10 seconds for the button to be attached to the DOM and visible.
-        # This handles the asynchronous loading of the Usercentrics script.
-        logger.info("Waiting for Usercentrics banner...")
-        deny_button.wait_for(state="visible", timeout=10000)
-
-        # 3. Click it once it appears
-        deny_button.click()
-        logger.success("Successfully clicked the 'Deny All' button.")
-
-    except TimeoutError:
-        # If 10 seconds pass and the button never shows up, it fails gracefully.
-        logger.error(
-            "Timeout: Banner did not appear within 10 seconds. "
-            "It may be disabled or already accepted."
-        )
-
-
-def scroll_to_top(page):
-    # Get the initial height of the page
-    page.evaluate(
-        """
-        var intervalID = setInterval(function () {
-            window.scrollBy(0, -window.innerHeight);
-        }, 200);
-        """
-    )
-    counter = 0
-    while True:
-        if page.evaluate("window.scrollY <= window.innerHeight"):
-            logger.info("I reached the top")
-            page.evaluate("clearInterval(intervalID)")
-            break
-        else:
-            counter += 1
-            logger.debug("Scrolling... counter={}", counter)
-            time.sleep(0.5)
 
 
 def get_store_configs(path: Path) -> list[StoreConfig]:
